+++ india_compliance/patches.txt
@@ -12,6 +12,7 @@ india_compliance.patches.post_install.set_gst_tax_type
 india_compliance.patches.post_install.set_gst_tax_type_in_journal_entry
 india_compliance.patches.post_install.update_company_gstin
 india_compliance.patches.post_install.update_custom_role_for_e_invoice_summary
+india_compliance.patches.post_install.migrate_fields_for_gstr3b
 india_compliance.patches.v14.remove_ecommerce_gstin_from_purchase_invoice
 india_compliance.patches.v14.set_sandbox_mode_in_gst_settings
 execute:from india_compliance.gst_india.setup import add_fields_to_item_variant_settings; add_fields_to_item_variant_settings()
@@ -62,4 +63,3 @@ execute:from india_compliance.gst_india.setup import set_default_print_settings;
 india_compliance.patches.v15.migrate_gstr1_log_to_returns_log
 india_compliance.patches.v15.update_action_for_gst_inward_supply
 india_compliance.patches.v15.set_default_for_new_gst_category_notification
-india_compliance.patches.v15.migrate_fields_for_gstr3b
